Add utility modules, API stubs, and session persistence
