+++ amalo/__init__.py
@@ -1,1 +1,9 @@
-"""Core calculation utilities."""
+"""Core calculation utilities.
+
+This module also exposes the package version for runtime display."""
+
+__all__ = ["__version__"]
+
+# Keep in sync with the version declared in ``setup.py``
+__version__ = "0.1.0"
+
