+++ app.py
@@ -1,209 +1,29 @@
-import json
 import streamlit as st
-from core.presets import (
-    PROGRAM_PRESETS,
-    CONV_MI_BANDS,
-    FHA_TABLES,
-    VA_TABLE,
-    USDA_TABLE,
-)
-from core.calculators import piti_components, dti, monthly_payment
-from core.models import W2
+
+from core.presets import PROGRAM_PRESETS
+from core.calculators import dti
+from core.state import load_session_state, save_session_state
 from ui.topbar import render_topbar
 from ui.cards_income import render_income_cards
 from ui.cards_debts import render_debt_cards
 from ui.bottombar import render_bottombar
 from ui.documents import render_document_checklist
+from ui.forms import render_w2_form, render_property_column
+from ui.sidebar import render_fee_sidebar
+from ui.dashboard import render_dashboard_view
+from ui.max_qualifiers import render_max_qualifiers_view
+
+# Re-export common rendering helpers for tests
+__all__ = [
+    "render_w2_form",
+    "render_property_column",
+    "render_fee_sidebar",
+    "render_dashboard_view",
+]
 
 
-# ---------------------------------------------------------------------------
-# Minimal W-2 form kept for test coverage
-# ---------------------------------------------------------------------------
+load_session_state()
 
-def render_w2_form():
-    st.session_state.setdefault("w2_rows", [])
-    if st.button("Add W2 Job", key="add_w2_job"):
-        st.session_state.w2_rows.append(W2().model_dump())
-    for idx, row in enumerate(st.session_state.w2_rows):
-        with st.expander(f"W2 #{idx+1}"):
-            items = list(row.items())
-            for i in range(0, len(items), 2):
-                cols = st.columns(2)
-                for col_idx, (field, val) in enumerate(items[i : i + 2]):
-                    with cols[col_idx]:
-                        st.markdown(f"**{field}**")
-                        st.caption(f"Enter {field}")
-                        st.text_input("", value=str(val), key=f"w2_{idx}_{field}")
-
-
-def fico_to_bucket(score):
-    """Map a numeric credit score to the preset FICO buckets."""
-    try:
-        s = float(score)
-    except (TypeError, ValueError):
-        return "760+"
-    if s >= 760:
-        return "760+"
-    if s >= 720:
-        return "720-759"
-    return "<720"
-
-
-def render_fee_sidebar():
-    """Sidebar with editable MI/MIP/funding fee tables."""
-    st.session_state.setdefault("conv_mi_table", CONV_MI_BANDS)
-    st.session_state.setdefault("fha_table", FHA_TABLES)
-    st.session_state.setdefault("va_table", VA_TABLE)
-    st.session_state.setdefault("usda_table", USDA_TABLE)
-
-    st.sidebar.header("MI / MIP / Guarantee")
-    conv_json = st.sidebar.text_area(
-        "Conventional MI Table",
-        value=json.dumps(st.session_state["conv_mi_table"], indent=2),
-    )
-    fha_json = st.sidebar.text_area(
-        "FHA MIP Table",
-        value=json.dumps(st.session_state["fha_table"], indent=2),
-    )
-    va_json = st.sidebar.text_area(
-        "VA Funding Fee Table",
-        value=json.dumps(st.session_state["va_table"], indent=2),
-    )
-    usda_json = st.sidebar.text_area(
-        "USDA Guarantee Fee Table",
-        value=json.dumps(st.session_state["usda_table"], indent=2),
-    )
-
-    try:
-        st.session_state["conv_mi_table"] = json.loads(conv_json)
-    except Exception:
-        pass
-    try:
-        st.session_state["fha_table"] = json.loads(fha_json)
-    except Exception:
-        pass
-    try:
-        st.session_state["va_table"] = json.loads(va_json)
-    except Exception:
-        pass
-    try:
-        st.session_state["usda_table"] = json.loads(usda_json)
-    except Exception:
-        pass
-
-
-# ---------------------------------------------------------------------------
-# Property / housing helpers
-# ---------------------------------------------------------------------------
-
-def render_property_column():
-    st.session_state.setdefault("housing", {})
-    h = st.session_state.housing
-    with st.expander("Payment & Housing"):
-        h["purchase_price"] = st.number_input("Purchase Price", value=float(h.get("purchase_price", 0.0)))
-        h["down_payment_amt"] = st.number_input("Down Payment", value=float(h.get("down_payment_amt", 0.0)))
-        h["rate_pct"] = st.number_input("Rate %", value=float(h.get("rate_pct", 0.0)))
-        h["term_years"] = st.number_input("Term (years)", value=float(h.get("term_years", 30)))
-        base_loan = h.get("purchase_price", 0.0) - h.get("down_payment_amt", 0.0)
-        pi_only = monthly_payment(
-            base_loan,
-            h.get("rate_pct", 0.0),
-            h.get("term_years", 30),
-        )
-        st.caption(f"Monthly P&I: ${pi_only:,.2f}")
-        h["tax_rate_pct"] = st.number_input(
-            "Tax Rate %",
-            value=float(h.get("tax_rate_pct", 0.0)),
-            help="Avg Florida property tax ~1% of purchase price",
-        )
-        h["hoi_rate_pct"] = st.number_input(
-            "HOI Rate %",
-            value=float(h.get("hoi_rate_pct", 0.0)),
-            help="Enter as % of purchase price (FL avg ~1%)",
-        )
-        h["hoi_annual"] = st.number_input(
-            "HOI Annual",
-            value=float(h.get("hoi_annual", 0.0)),
-            help="Annual homeowners insurance amount",
-        )
-        if h.get("hoi_rate_pct", 0.0) > 0:
-            h["hoi_annual"] = h.get("purchase_price", 0.0) * h["hoi_rate_pct"] / 100
-            st.caption(f"Calculated HOI Annual: ${h['hoi_annual']:,.2f}")
-        h["hoa_monthly"] = st.number_input(
-            "HOA Monthly",
-            value=float(h.get("hoa_monthly", 0.0)),
-            help="Florida HOA averages ~$250/mo",
-        )
-        h["finance_upfront"] = st.checkbox("Finance Upfront Fees", value=bool(h.get("finance_upfront", True)))
-        h["credit_score"] = st.number_input(
-            "Credit Score", value=float(h.get("credit_score", 760))
-        )
-        h["first_use_va"] = st.checkbox(
-            "First Use VA", value=bool(h.get("first_use_va", True))
-        )
-    base_loan = h.get("purchase_price", 0.0) - h.get("down_payment_amt", 0.0)
-    comps = piti_components(
-        st.session_state.get("program_name", "Conventional"),
-        h.get("purchase_price", 0.0),
-        base_loan,
-        h.get("rate_pct", 0.0),
-        h.get("term_years", 30),
-        h.get("tax_rate_pct", 0.0),
-        h.get("hoi_annual", 0.0),
-        h.get("hoa_monthly", 0.0),
-        st.session_state.get("conv_mi_table", CONV_MI_BANDS),
-        st.session_state.get("fha_table", FHA_TABLES),
-        st.session_state.get("va_table", VA_TABLE),
-        st.session_state.get("usda_table", USDA_TABLE),
-        h.get("finance_upfront", True),
-        h.get("first_use_va", True),
-        fico_to_bucket(h.get("credit_score")),
-    )
-    st.session_state["housing_calc"] = comps
-    program = st.session_state.get("program_name", "Conventional")
-    if h.get("finance_upfront", True) and program in ("FHA", "VA", "USDA"):
-        st.caption(
-            f"Base Loan: ${base_loan:,.0f} • Adjusted Loan: ${comps['adjusted_loan']:,.0f}"
-        )
-    else:
-        st.caption(f"Base Loan: ${base_loan:,.0f}")
-    st.caption(f"LTV: {comps['ltv']*100:.2f}%")
-    st.caption(f"PITIA: ${comps['total']:,.2f}")
-    return comps
-
-
-# ---------------------------------------------------------------------------
-# Dashboard view
-# ---------------------------------------------------------------------------
-
-def render_dashboard_view(summary):
-    st.header("Dashboard")
-    cols = st.columns(4)
-    cols[0].metric("Total Income", f"${summary['total_income']:,.2f}")
-    cols[1].metric("PITIA", f"${summary['pitia']:,.2f}")
-    cols[2].metric("FE DTI", f"{summary['fe_dti']*100:.2f}%")
-    cols[3].metric("BE DTI", f"{summary['be_dti']*100:.2f}%")
-    from core.rules import evaluate_rules
-
-    state = {
-        "total_income": summary["total_income"],
-        "FE": summary["fe_dti"],
-        "BE": summary["be_dti"],
-        "target_FE": st.session_state["program_targets"]["fe_target"],
-        "target_BE": st.session_state["program_targets"]["be_target"],
-    }
-    for r in evaluate_rules(state):
-        if r.severity == "critical":
-            st.error(f"[{r.code}] {r.message}")
-        elif r.severity == "warn":
-            st.warning(f"[{r.code}] {r.message}")
-        else:
-            st.info(f"[{r.code}] {r.message}")
-
-
-# ---------------------------------------------------------------------------
-# Main entry point
-# ---------------------------------------------------------------------------
 
 def main():
     st.set_page_config(layout="wide")
@@ -218,11 +38,6 @@ def main():
     st.session_state.setdefault("ui_prefs", {"show_bottom_bar": False, "language": "en"})
     render_fee_sidebar()
 
-    # Render the top bar and capture current program/target selections. The
-    # selectbox in ``render_topbar`` already manages ``program_name`` via
-    # Streamlit's ``session_state``. Attempting to assign to the same key again
-    # triggers a ``StreamlitAPIException`` in recent versions of Streamlit.
-    # Rely on the widget-managed value instead of overwriting it here.
     view_mode, targets, _program = render_topbar()
     st.session_state["program_targets"] = targets
     if view_mode == "data_entry":
@@ -267,8 +82,6 @@ def main():
         }
         render_dashboard_view(summary)
     else:
-        from ui.max_qualifiers import render_max_qualifiers_view
-
         render_max_qualifiers_view()
 
     st.sidebar.checkbox(
@@ -295,7 +108,9 @@ def main():
         """,
         unsafe_allow_html=True,
     )
+    save_session_state()
 
 
 if __name__ == "__main__":
     main()
+
